Extract some utils to utils.py
